Added filter decorators for tests
When you pass a noun filter phrase to a test, the filter decorator will
translate it into a list of nouns and pass that back to the method.

Stored in a separate module so these can also be imported into actions.
Yes, I know I'm getting ahead of myself, but I know I'll be using these
in actions and besides, I couldn't resist.